Remove topic_coherence submodule, create new file with all additional functions.
