+++ gensim/test/test_direct_confirmation.py
@@ -11,9 +11,7 @@ Automated tests for direct confirmation measures in the direct_confirmation_meas
 import logging
 import unittest
 from collections import namedtuple
-
-from gensim.topic_coherence import direct_confirmation_measure
-from gensim.topic_coherence import text_analysis
+import gensim.models.coherence_utils
 
 
 class TestDirectConfirmationMeasure(unittest.TestCase):
@@ -27,45 +25,45 @@ class TestDirectConfirmationMeasure(unittest.TestCase):
         id2token = {1: 'test', 2: 'doc'}
         token2id = {v: k for k, v in id2token.items()}
         dictionary = namedtuple('Dictionary', 'token2id, id2token')(token2id, id2token)
-        self.accumulator = text_analysis.InvertedIndexAccumulator({1, 2}, dictionary)
+        self.accumulator = gensim.models.coherence_utils.InvertedIndexAccumulator({1, 2}, dictionary)
         self.accumulator._inverted_index = {0: {2, 3, 4}, 1: {3, 5}}
         self.accumulator._num_docs = self.num_docs
 
     def testLogConditionalProbability(self):
         """Test log_conditional_probability()"""
-        obtained = direct_confirmation_measure.log_conditional_probability(
+        obtained = gensim.models.coherence_utils.log_conditional_probability(
             self.segmentation, self.accumulator)[0]
         # Answer should be ~ ln(1 / 2) = -0.693147181
         expected = -0.693147181
         self.assertAlmostEqual(expected, obtained)
 
-        mean, std = direct_confirmation_measure.log_conditional_probability(
+        mean, std = gensim.models.coherence_utils.log_conditional_probability(
             self.segmentation, self.accumulator, with_std=True)[0]
         self.assertAlmostEqual(expected, mean)
         self.assertEqual(0.0, std)
 
     def testLogRatioMeasure(self):
         """Test log_ratio_measure()"""
-        obtained = direct_confirmation_measure.log_ratio_measure(
+        obtained = gensim.models.coherence_utils.log_ratio_measure(
             self.segmentation, self.accumulator)[0]
         # Answer should be ~ ln{(1 / 5) / [(3 / 5) * (2 / 5)]} = -0.182321557
         expected = -0.182321557
         self.assertAlmostEqual(expected, obtained)
 
-        mean, std = direct_confirmation_measure.log_ratio_measure(
+        mean, std = gensim.models.coherence_utils.log_ratio_measure(
             self.segmentation, self.accumulator, with_std=True)[0]
         self.assertAlmostEqual(expected, mean)
         self.assertEqual(0.0, std)
 
     def testNormalizedLogRatioMeasure(self):
         """Test normalized_log_ratio_measure()"""
-        obtained = direct_confirmation_measure.log_ratio_measure(
+        obtained = gensim.models.coherence_utils.log_ratio_measure(
             self.segmentation, self.accumulator, normalize=True)[0]
         # Answer should be ~ -0.182321557 / -ln(1 / 5) = -0.113282753
         expected = -0.113282753
         self.assertAlmostEqual(expected, obtained)
 
-        mean, std = direct_confirmation_measure.log_ratio_measure(
+        mean, std = gensim.models.coherence_utils.log_ratio_measure(
             self.segmentation, self.accumulator, normalize=True, with_std=True)[0]
         self.assertAlmostEqual(expected, mean)
         self.assertEqual(0.0, std)
