--- gensim/topic_coherence/__init__.py
@@ -1,4 +1,0 @@
-"""
-This package contains implementation of the individual components of
-the topic coherence pipeline.
-"""
